Inital import of the 2014 design system CSS
Add initial layer of CSS for the 2014 design system. This includes the
base, gel, generic and helper folders, but exluces the object and module
folders. We should add the contents of those folders in as and when we
need them, as their folder structure shall be different to what it was
previously as we're now doing the Atomic Design and placing css and
templates in the same folder thing.