Inital work on creating an organism presenter and translations
An example of the presenter ideology. The controller gives domain
objects to the view, which then uses a custom Twig extension to pass
that into a Presenter object and render its template. Much of the logic
within the base Presenter and Factory was taken from Amen.

Translations end up recreating the translate class twice (once in DI
container building and once in BaseController) if a branding that
uses anything but the default (en_GB) language. This is slightly
inefficient but acceptable for now as the vast majority of things are in
en_GB.