Pass in Translate to presenters that need it
Now that the ds2013() twig function exists we don't need to do weird
things with passing the PresenterFactory into every presenter because
the twig template deals with the creation of any dependant presenters
(e.g. rending an image presenter within a programme presenter).
No presenters shall need the Factory passed in as an argument.

However we still sometimes want to use the translation function within
some Presenters. So create a TranslatablePresenterTrait for these
presenters to use. These presenters should pass an instance of Translate
in as the first argument (not a hard and fast rule but seems like a
sensible convention).