Presenter Template tests
A framework for creating tests of presenter templates for when we want
to test the logic inside them. Most of the time we should be able to
test the Presenter directly, but there are some cases where we want to
assert on the output of view.

Test broadcast to see some examples.