Redirect to programme options location if required
Add ProgrammeOptionsRedirectHttpException that returns a redirect to the
given location.

Throw that exception when we get an entity in the
ContextEntityByPidValueResolver and FindByPidRouterSubscriber.

This is done in both places as if the FindByPidRouterSubscriber adds a
domain object to the request attributes then the
ContextEntityByPidValueResolver is not fired, because the
RequestAttributeValueResolver resolves the controller arguments first.