Update to Symfony 3.4.1 and CsaGuzzleBundle 3.0.1
The ability to modify the container before using it in tests was
undeprecated so we don't need to tag the tests that used that
pattern as legacy anymore.

Liip\FunctionalTestBundle currently throws deprecations, so overwrite
the method that caused them a version that doesn't check deprecated
features.

Updating the framework files (e.g. AppKernel, the frontcontrollers etc)
to use their latest v3.4 versions will be done in a later commit, as
part of moving the folder structure around to follow Flex conventions.

We can't update to Symfony 4.0 yet as Liip\FunctionalTestBundle doesn't
support it yet.