Define the :pid/episodes/player route correctly
Define the :pid/episodes/player route in routing_3rd_party.yml with all
the other routes that are currently served by the PAL. Name it similar
to the other existing routes and set the required parameters on it - Add
a _controller, and set the scheme for the route correctly to avoid a
redirect.

Add a check for a valid _controller to the MetricsManager so a single
badly configured route (i.e. lacking a _controller) does not cause every
page in the app to throw a 500 error.