Move to a Flex style config layout
Use Flex as a tool for config management, update frontcontroller and
console based on the new flex recipes (yay one index.php). Update config
directory to use the flex style

As environment variables have caused us (Bjorn) pain in the past
continue to use a generated parameters.yaml for such config instead of
using environment variables. Thus there is no .env.dist (and a generated
.env), but parameters.yaml.dist and a generated parameters.yaml.

Comment out the mention of dotenv in index.php and bin/console to make
it clear that we know it should be there but we're choosing not to use
it