Custom conifg loading for derivative envs
Keeping the standard config loading mechanism results in a lot of
boilerplate and potential for doing things out of order.

Introduce the concept of derivative environments in the Kernel's config
loader so that prod_int automatically loads in prod's config beforehand
and prod_int_fixture automatically loads in prod's config, prod_int's
config and the fixture config.

This also fixes the config load ordering for these derived environments
so the load ordering between the packages folder and services_*.yaml
works in the way you'd expect.