Remove deprecations by removing overrides for render() and redirect()
These methods are marked @final in Symfony 3.4 so will trigger
deprecations if you try and override them in the BaseController. So
instead create alternative functions that shall pick up on the caching
headers.

Instead of render() use renderWithChrome() or renderWithoutChrome()
to ensure caching headers are set.

Instead of redirect() / redirectToRoute(), use cachedRedirect() and
cachedRedirectToRoute() to ensure caching headers are set.